// src/lib/api/generated/model/getFoodByBarcode200.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Calorie Compass AWS API
 * Minimal authenticated API contract for Cognito + API Gateway + Lambda.
 * OpenAPI spec version: 0.1.0
 */
import type { BarcodeLookupItem } from './barcodeLookupItem';
import type { GetFoodByBarcode200Status } from './getFoodByBarcode200Status';

export interface GetFoodByBarcode200 {
  status: GetFoodByBarcode200Status;
  item: BarcodeLookupItem;
}